ai-evaluation-service/requirements.txt
azure-ai-projects
azure-identity
python-dotenv>=0.19.0
Flask>=2.0.0
flask-cors>=4.0.0
